bitmap

+++ include/sys/types.h
@@ -2,13 +2,13 @@
 #define __TYPES_H__
 
 typedef unsigned char   uint8_t;
-typedef char            int8_t;
+typedef signed char     int8_t;
 typedef unsigned short  uint16_t;
-typedef short           int16_t;
-typedef unsigned long   uint32_t;
-typedef long            int32_t;
+typedef signed short    int16_t;
+typedef unsigned int    uint32_t;
+typedef signed int      int32_t;
 typedef unsigned long long uint64_t;
-typedef long long       int64_t;
+typedef signed long long   int64_t;
 
 typedef int32_t         error_t;
